requirements.txt
flask==2.3.3
python-dotenv
twilio
google-generativeai
requests
mcp
google-genai
redis>=5.0.0
rq>=1.16.0
gunicorn>=22.0.0
waitress>=3.0.0